Include linux/dma-mapping.h in linux/dmaengine.h
This reverses the effect of b7f080cfe223b3b7424872639d153695615a9255 (net:
remove mm.h inclusion from netdevice.h) while leaving almost all of the
code changes in place, in particular the creation of linux/dma-diretion.h
which is a dependency of subsequent patches that will be backported.

The motivation for this it to unbreak multiple drivers.
This patch cam be reverted if backports to all relevant drivers
are made.

Signed-off-by: Simon Horman <horms@verge.net.au>