gpio: Emma Mobile GPIO driver
This patch adds a GPIO driver for the Emma Mobile line
of SoCs. The driver is designed to be reusable between
multiple SoCs that share the same basic building block,
but so far it has only been used on Emma Mobile EV2.

Each driver instance handles 32 GPIOs with individually
maskable IRQs. The driver operates on two I/O memory
ranges and the 32 GPIOs are hooked up to two interrupts.

In the case of Emma Mobile EV2 this GPIO building block
is used as main external interrupt controller hooking up
159 GPIOS as 159 interrupts via 5 driver instances and
10 interrupts to the GIC and the Cortex-A9 Dual.

This version of the driver gets all configuration via
struct resource and platform data, but DT bindings are
currently in development and such support will be sent
as a individual feature patch in the near future.

Signed-off-by: Magnus Damm <damm@opensource.se>

N.B: This driver has been merged upsteram
as commit a07e103ef08c6907d695a06467d7ee950796fccf.
However, this code is an earlier version of
the driver that doesn't make use of infrastructure
that isn't present in Linux 3.0.

Signed-off-by: Simon Horman <horms@verge.net.au>